requirements.txt
# YOLOv8
ultralytics==8.0.234

# Image processing
opencv-python==4.8.1.78
Pillow==10.1.0

albumentations>=1.3.0

# Data handling
numpy==1.24.3
pandas==2.0.3

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
tqdm==4.66.1
pyyaml==6.0.1

# For GPU support (optional)
torch==2.1.0
torchvision==0.16.0

gradio>=3.50.0
